requirements.txt
fastapi
uvicorn
pdfplumber
pytesseract
Pillow
pydantic
python-multipart
pytest
httpx